+++ src/crate/theme/rtd/conf/sql_99.py
@@ -19,17 +19,23 @@
 # with Crate these terms will supersede the license and you may use the
 # software solely pursuant to the terms of the relevant commercial agreement.
 
+
 from crate.theme.rtd.conf import *
 
-project = u'SQL 99'
+# If you update the `project` value here, you must update it in the
+# `src/crate/theme/rtd/crate/sidebartoc.html` file or else Sphinx will not
+# expand the sidebar TOC for this project.
+project = u"SQL 99"
 html_title = project
 
-url_path = 'docs/sql-99'
+url_path = "docs/sql-99"
 
 # For sitemap extension
-html_baseurl = 'https://crate.io/%s/' % url_path
+html_baseurl = "https://crate.io/%s/" % url_path
 
 # For rel="canonical" links
-html_theme_options.update({
-    'canonical_url_path': '%s/en/latest/' % url_path,
-})
+html_theme_options.update(
+    {
+        "canonical_url_path": "%s/en/latest/" % url_path,
+    }
+)